+++ test/fixtures/load/src/main.js
@@ -1,0 +1,3 @@
+import msg1 from './msg'
+
+console.log(msg1)