+++ test/fixtures/load/src/msg.js
@@ -1,0 +1,1 @@
+export default 'it is a msg'